+++ src/main/java/org/zerock/puppyrun/auth/controller/response/SignInResponse.java
@@ -1,9 +1,9 @@
-package org.zerock.puppyrun.member.DTO;
+package org.zerock.puppyrun.auth.controller.response;
 
 import lombok.Builder;
 
 @Builder
-public record TokenDTO(
+public record SignInResponse(
         String accessToken,
         String refreshToken
 ) {
